backend/requirements.txt
fastapi>=0.110.0
spacy>=3.8.0
spacy-lookups-data>=1.0.5
uvicorn>=0.30.0
pydantic>=2.7.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
python-multipart>=0.0.9
pypdf>=5.0.0
python-docx>=1.1.0